m: Move reaper-related code to another module
